requirements.txt
camelot-py
numpy
pandas
coloredlogs
dateutil
click
